umrah: fix HTTP 500 on invoice view + PDF download (empty tax string)
Umrah bookings store `tax = ''` (empty string, not NULL) — the checkout insert
never sets it. In PHP 8, arithmetic / number_format on an empty string is a fatal
TypeError, which was 500-ing two owner-facing paths (anonymous callers redirect
out before hitting them, so this only bit legitimate owners — and every umrah
booking):

1. app/routes/umrah/invoiceRoutes.php:92 — the invoice-viewed webhook computed
   `total_amount => price_original + price_markup + tax`. `?? 0` doesn't catch
   '' (only NULL), so `float + ''` threw "Unsupported operand types" → 500 on
   GET /invoice/umrah/{id} for the owner. Cast each component to float.

2. app/views/notifications/emails/umrah/booking_pdf.php:434 —
   `number_format($booking['tax'], 2)` threw "number_format(): Argument #1 must
   be of type float, string given" inside the PDF template include, so
   GENERATE_BOOKING_PDF returned false → GET /api/umrah/booking/download-invoice
   500'd AND resend-invoice couldn't attach a PDF. Cast to float with a null-safe
   default.

Verified over real HTTP: owner invoice view 200, PDF download 200
(application/pdf, 35KB, 1 page), resend 200 (email sent) / 400 without CSRF.
IDOR unchanged and confirmed: anonymous → view 302→login, PDF 403; a different
logged-in user → view 302, PDF 403. Test data cleaned up.

+++ app/routes/umrah/invoiceRoutes.php
@@ -89,7 +89,11 @@ $router->get('/invoice/umrah/([a-zA-Z0-9]+)', function ($invoiceId) use ($SECURE
         'user_id' => $booking['user_id'],
         'payment_status' => $booking['payment_status'],
         'booking_status' => $booking['booking_status'],
-        'total_amount' => ($booking['price_original'] ?? 0) + ($booking['price_markup'] ?? 0) + ($booking['tax'] ?? 0),
+        // Cast each component to float: these decimal columns come back as strings
+        // and `tax` is often an empty string '' (not NULL, so ?? 0 doesn't catch
+        // it) on umrah bookings — `float + ''` is a fatal TypeError in PHP 8, which
+        // was 500-ing the invoice view for the OWNER (anon callers redirect earlier).
+        'total_amount' => (float) ($booking['price_original'] ?? 0) + (float) ($booking['price_markup'] ?? 0) + (float) ($booking['tax'] ?? 0),
         'timestamp' => date('Y-m-d H:i:s'),
         'ip_address' => $_SERVER['REMOTE_ADDR'] ?? '0.0.0.0',
         'user_agent' => $_SERVER['HTTP_USER_AGENT'] ?? 'Unknown'
